# backend/api/api-spec.yml
openapi: 3.1.0
info:
  title: Tasktify API
  version: 1.0.0
  description: |
    Task management API with Post-Quantum Cryptography (PQC) JWT authentication.
    Supports multiple signing algorithms: Falcon-512, ML-DSA (44/65/87), SLH-DSA,
    and classical algorithms (ES256, RS256, HS256, EdDSA).
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/profile:
    get:
      tags: [Profile]
      summary: Get authenticated user profile
      operationId: getProfile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/User"
              example:
                status: 200
                data:
                  id: 550e8400-e29b-41d4-a716-446655440000
                  name: John Doe
                  email: john@example.com
                  created_at: 1714204800000
                  updated_at: 1714204800000
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        PQC-signed JWT token obtained from `POST /api/auth/signin`.
        Token header contains the algorithm identifier (e.g. `Falcon-512`, `ML-DSA-44`).
...
  responses:
...
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: 401
            message: unauthorized
...
  schemas:
    Response:
      type: object
      properties:
        status:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          description: Response payload (type varies by endpoint)
        paging:
          $ref: "#/components/schemas/PageMetadata"
        error:
          description: Error details (present only on failure)
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        error:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
    PageMetadata:
      type: object
      properties:
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 10
        total_item:
          type: integer
          format: int64
          example: 42
        total_page:
          type: integer
          format: int64
          example: 5
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john@example.com
        created_at:
          type: integer
          format: int64
          description: Unix timestamp in milliseconds
          example: 1714204800000
        updated_at:
          type: integer
          format: int64
          description: Unix timestamp in milliseconds
          example: 1714204800000
...